+++ docs/changelog.rst
@@ -1,9 +1,15 @@
 Changelog
 =========
 
-0.7.0
+0.8.0
 -----
 - added support for django-rest-framework-jwt
+0.7.0
+-----
+- Wrapped API returned strings in ugettext_lazy
+- Fixed not using ``get_username`` which caused issues when using custom user model without username field
+- Django 1.9 support
+- Added ``TwitterLoginSerializer``
 
 0.6.0
 -----